requirements.txt
streamlit
pinecone-client
openai
pdfminer.six
requests
beautifulsoup4
python-dotenv
numpy
reportlab